refactor(agent): extract tool execution helpers from engine_run.go
- Extract executeToolCalls into: initToolProgress, execOneTool,
  updateToolResultProgress, updateToolResultLine, dispatchToolCalls
  (with 3 strategy methods), executeSubAgentOps, snapshotCompletedIteration
- Extract maybeMaskObservations from maybeCompress
- Extract setWaitingUser from processToolResults
- Result: largest function reduced from 291 to ~5 lines (body)
- All functions now under 103 lines
